feat(payment): add queue-acceptance flow with sliding-window PG throttling

+++ src/main/java/com/mudosa/musinsa/payment/application/dto/response/PaymentQueueResponse.java
@@ -7,8 +7,8 @@ import lombok.Getter;
 @Builder
 public class PaymentQueueResponse {
     private Long ticketId;
-    private long queuePosition;
-    private int estimatedWaitSeconds;
+    private Long queuePosition;
+    private Long estimatedWaitSeconds;
     private String status;
     private String orderNo;
 }
